+++ library/Swagger/Annotations/Operation.php
@@ -100,6 +100,11 @@ class Operation extends AbstractAnnotation
     public $consumes;
 
     /**
+     * @var Authorizations
+     */
+    public $authorizations;
+
+    /**
      * Undocumented
      * @var bool
      */
@@ -111,6 +116,7 @@ class Operation extends AbstractAnnotation
         '\Swagger\Annotations\Produces' => 'produces[]',
         '\Swagger\Annotations\Consumes' => 'consumes[]',
         '\Swagger\Annotations\Items' => 'items',
+        '\Swagger\Annotations\Authorizations' => 'authorizations',
     );
 
     /**