+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 Changelog
 ----------
 
+## 0.9.x
+
+* Implemented @SWG\Info & @SWG\Authorizations which can be used to augment the api-docs.json with metadata.
+
 ## 0.9.1
 
 * Generate ".{format}" instead of ".json" suffix when using the commandline.
@@ -12,7 +16,7 @@ Changelog
 
 * New processing architecture allowing swagger-php to act on other annotations (https://github.com/zircote/swagger-php/pull/121)
 * Improved model referencing: $ref allowed with complex modelnames (https://github.com/zircote/swagger-php/pull/122 and https://github.com/zircote/swagger-php/pull/123)
-* Updated depedencies creating a 50% smaller swagger.phar
+* Updated dependencies creating a 50% smaller swagger.phar
 
 ## 0.8.3